requirements.txt
flask==2.0.1
pyyaml==5.4.1
# Add any other dependencies required by the Open Interpreter